pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "memento"
version = "0.1.0"
description = "Reusable documentation portal with GitHub issues viewer"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "authlib>=1.3",
    "psycopg2-binary>=2.9",
    "httpx>=0.27",
    "pyjwt[crypto]>=2.8",
    "pyyaml>=6.0",
    "markdown>=3.5",
    "pymdown-extensions>=10.0",
    "python-dotenv>=1.0",
    "gunicorn>=22.0",
]

[tool.setuptools.packages.find]
include = ["memento*"]

[project.scripts]
memento = "memento.app:main"
